315、商城业务-秒杀服务-秒杀商品上架-2

+++ gulimall-seckill/src/main/java/com/atguigu/gulimall/seckill/vo/SeckillSkuVo.java
@@ -26,7 +26,7 @@ public class SeckillSkuVo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private int seckillCount;
     /**
      * 每人限购数量
      */
